+++ python/GafferSceneTest/__init__.py
@@ -144,6 +144,7 @@ from CopyPrimitiveVariablesTest import CopyPrimitiveVariablesTest
 from SpreadsheetTest import SpreadsheetTest
 from MergeScenesTest import MergeScenesTest
 from ShuffleAttributesTest import ShuffleAttributesTest
+from ShufflePrimitiveVariablesTest import ShufflePrimitiveVariablesTest
 
 from IECoreScenePreviewTest import *
 from IECoreGLPreviewTest import *