+++ python/GafferSceneUI/__init__.py
@@ -156,6 +156,7 @@ import DeformerUI
 import CopyPrimitiveVariablesUI
 import MergeScenesUI
 import ShuffleAttributesUI
+import ShufflePrimitiveVariablesUI
 
 # then all the PathPreviewWidgets. note that the order
 # of import controls the order of display.